# target/classes/swagger/swagger.yaml
swagger: '2.0'
info:
  description: REST service to generate and parse QR barcodes.
  version: '1.1'
  title: Swiss QR Barcodes
schemes:
  - https
  - http
basePath: /service
...
paths:
...
  /basic/barcode:
    get:
      tags:
        - BasicService
      summary: Returns a barcode image for the indicated content string.
      description: >-
        Returns a barcode image for the indicated content string. For Swiss QRs
        we only render the swiss cross when there are no errors
      operationId: getBarcode
      produces:
        - image/png
        - image/jpeg
        - image/gif
      parameters:
        - in: header
          name: licenseKey
          description: purchased license key
          type: string
        - name: content
          in: query
          type: string
          description: Content string that the barcode represents
        - name: dimension
          in: query
          default: 46
          type: number
          description: Height and width in mm
        - name: errorCorrectionLevel
          in: query
          default: M
          type: string
          enum:
            - H
            - L
            - M
            - Q
          description: >-
            Error correction level which defines the quality in the following
            sequence: Low->Mediaum->Quality->High
      responses:
        '200':
          description: image file with barcode
          schema:
            type: file
...